The problem occurs when using cg_goto() on files that are “linked”.  As I am sure most MLL users do, we sometimes use cg_goto() to navigate to a node, using the return value CG_NODE_NOT_FOUND to detect the absence of the given node.  When you try to navigate to an undefined node within a linked node, you get STDOUT message “#### BAD ID [ 1771]” – if the node is not linked, you don’t get the message.
I’ve traced it to an errant HDF5 “valid ID” check in the CGNS sources.  The offending statement (highlighted in snippet below) is on line 1843 of src/adfh/ADFH.c.  I believe the statement should just be removed, as it appears normal to have sid==-1 when the node is “not found”, and the MLL will ultimately return CG_NODE_NOT_FOUND.  If you look at the logic for the “non-link” get_node_ID, just below this logic, you’ll see that the ADFH_CHECK_HID(sid) statement is absent, which is what led me to believe this #1 statement should be deleted.

  else if (is_link(hpid)) {
    hid_t lid = open_link(hpid, err);
    ADFH_DEBUG(("<ADFH_Get_Node_ID open_link [%d]",lid));
    if (lid < 0) return;
    sid = H5Gopen2(lid, name, H5P_DEFAULT);
    ADFH_CHECK_HID(sid); #1
    H5Gclose(lid);
    if(sid < 0)
    {
      set_error(ADFH_ERR_GOPEN, err);
    }
  }

Reported by Mark Renzulli.