Patch from mailing list:
Hi,

i tested the CGNS-v3.3.0-alpha.2 to write fields in parallel providing NULL as parameter for the data pointer in all but one process. This led to an error message from the HDF5 library.

When writing data in parallel the readwrite_data_parallel method tests if the current process intends to write data by checking if the data pointer is NULL. In every method calling readwrite_data_parallel the data pointer is never NULL, it always points to a cg_rw_t struct. This leads to the observed error message.

When the user provides NULL as pointer for the cgp_*_read/write_dataset APIs, the NULL pointer is stored in the cg_rw_t struct in .u.wbuf or .u.rbuf. Therefore i changed the test in readwrite_data_parallel to check if the data.u.wbuf or data.u.rbuf pointer is null in the read, respectively write case.

This fixed the issue for me. I appended a git diff of the adapted pcgnslib.c file.

CGNS-51