Add prototype for cgp_parent_data_write
A first shot at an implementation for cgp_parent_data_write.  It works in the testing I have done on my local project, but not sure if the "style" is correct.  It currently uses an MPI_Allreduce call to determine the total size of the ParentElements and ParentElementsPosition nodes which are created during the function. It might be cleaner to have two functions -- one to create the nodes and one to write the data to the nodes, but this implementation follows the behavior of the serial version.

Let me know what you think and I will modify as needed.