Modify CHECK_FORTRAN_FEATURE macro
Change the extension of the file created by the CHECK_FORTRAN_FEATURE macro from .f90 (lowercase) to .F90 (uppercase) so that the fortran preprocessor will be run on the file.  This was causing a false negative on xlf compiler for the CONCAT check since the compilation was failing due to the preprocessor not being run instead of failing due to an incorrect concatenation macro.

For tests that do not need preprocessor to be run, this should still be ok and not cause any problems.