If processor has no data, set end = start = 0
If a processor has no data to write, then the start and end values passed in are sometimes arbitrary from calling code.  For example end > start or end == start.  Depending on the values, there can be logic errors in the following code.  To avoid this, set start = end = 0 if a processor has no data to output (section->parelem == NULL).