Improve error messages, fix error check (fixed)
Improved the error messages.  The original message for checking for valid range would output something that didn't make sense or at least wasn't helpful.  Split this into two checks and output info that helps locate problem.

Added `connectname` to error messages to show which connection had issues.

The transform check was testing `abs(transform[i]) < 0` which isn't possible so that can be removed.  Note that transform[i] can be equal to zero in the null dimension which wasn't recognized in the original pull request.

This was originally submitted as #62 which was applied and then reverted by #63.  Issues identified in #63 have been addressed in this revision.