Make static mpi-related variables consistent (#139)
* Make static mpi-related variables consistent

Currently, it is possible for the `pcg_mpi_comm` to be inconsistent with `pcg_mpi_comm_rank` and `pcg_mpi_comm_size`.  
The latter two values are set during a call to `cgp_open,` and the first is set during a call to `cgp_mpi_comm.`  If the user is using both parallel and serial file access during the same run, they may call `cgp_mpi_comm` multiple times, and the values of `pcg_mpi_comm_rank` and `pcg_mpi_comm_size` will only be consistent with the communicator in `pcg_mpi_comm` if the user calls `cgp_open.`  

With the changes here, all three values (and the `ParallelMPICommunicator`) will always be consistent.  Also, if the user does not call `cgp_mpi_comm()` before calling `cgp_open(),` this will be detected, and everything will be consistent.

* Set communicator to MPI_COMM_NULL

* Initialize communicator to MPI_COMM_NULL