Change shared build rpath test symbol
I think that the correct symbol to test should be `CGNS_BUILD_SHARED` instead of `CGNS_USE_SHARED`

With the test as it currently exists, my builds are not getting the correct rpath during the install of the tools (cgnsidff, cgnslist, ...), but if I use `CGNS_BUILD_SHARED`, then everything seems to work correctly and i get the correct rpath on the tools.  The `make install` step also shows that CMake is adding the runtime path to the installed tools, but I don't get the message with the current symbol in the test.  (The message is `Set runtime path of ...` during the install phase)