Handle integer size correctly
I think this fixes the problem of an integer mismatch between the size of `cgsize_t` and the size of the data stored in the `Elements` array in the CGNS file.  The data will be read from the file and stored in the `elements` array which is declared as `cgsize_t *elements`, so I think that should be the `type` that is passed do `readwrite_data_parallel`.  If `cgsize_t` is different than the size of the data in the file, then I think HDF5 will do the conversion.   I've tested this on the 4 possible settings (cgsize_t = 4 or 8, size of `Elements` on disk = I4 or I8) and they all work correctly with the fix.  Without the fix, I get a failure whenever the sizes don't match for a parallel read.